int/utils/indent: renames package to simplify calls after it is left with only one symbol.
