+++ internal/evolution/evaluation/evaluator_test.go
@@ -3,12 +3,13 @@ package evaluation
 import (
 	"fmt"
 	"sort"
+	"testing"
+
 	"tde/internal/evolution/evaluation/inject"
 	"tde/internal/evolution/evaluation/list"
 	"tde/internal/evolution/evaluation/slotmgr"
 	"tde/internal/evolution/models"
-	"tde/internal/utilities/strw"
-	"testing"
+	"tde/internal/utilities/indent"
 )
 
 func prepareEvaluator(ctx *models.Context) (*Evaluator, error) {
@@ -57,7 +58,7 @@ func Test_Pipeline(t *testing.T) {
 
 				if subject.Fitness.Layer() != layer {
 					t.Errorf("assert, layer mistmatch: got=%q (%s/%d, fitness=%.3f)\n%s",
-						subject.Fitness.Layer(), layer, i, subject.Fitness.Flat(), strw.IndentLines(string(subject.Code), 4))
+						subject.Fitness.Layer(), layer, i, subject.Fitness.Flat(), indent.Lines(string(subject.Code), 4))
 				}
 
 				sorted = append(sorted, subject)
