+++ internal/utilities/indent/strings.go
@@ -1,4 +1,4 @@
-package strw
+package indent
 
 import (
 	"strings"
@@ -6,7 +6,7 @@ import (
 	"tde/internal/utilities/functional"
 )
 
-func IndentLines(str string, indent int) string {
+func Lines(str string, indent int) string {
 	indentation := strings.Repeat(" ", indent)
 	return strings.Join(functional.Map(strings.Split(str, "\n"), func(i int, line string) string {
 		return indentation + line
